Changing the version detection heuristic.
trace-summary now looks at the number of columns it finds to determine
whether it's v1 or v2. Previously, it looked for '#', but I'm thinking
that might get stripped out if somebody pipes the conn.log through
some other tool first. And if it is missing, trace-summary would
determine it's v1, which is an unfortuante default going forward.

The new heuristic isn't perfect either, but it should work quite well
and it fails towards favoring v2.