Replace print statements and fix output for -R option
Replaced all print statements (does not exist in Python 3).  Also fixed
timestamps in output files when running Python 2 and using the "-R" option
so that they are no longer rounded unnecessarily when being converted to
strings, which was causing some loss of precision (noticed this
issue because Python 3 doesn't do this).  Updated test baselines.