Change trace-summary so that it reads the log header lines
Changed trace-summary so that it tries to extract the field separator,
unset field string, and field positions from the log header metadata lines.
Upon failure to extract this data, it will print a warning and continue.
The hard-coded default values are used if there is no metadata or if
extraction fails.

Also changed the version auto-detection logic in the case where a log
file has a metadata line.  In this case, trace-summary assumes it is
a version 2 log file (previously, trace-summary just ignored all
metadata lines).