Fix a minor bug when using Python 3
In the gethostbyaddr function, if the child process takes too long to
return a DNS lookup, it is killed and then Python 3 returns from the
call to waitpid(), but then trace-summary incorrectly proceeds to read
from the pipe.  The fix is to always check the exit status before
attempting to read the pipe.  This bug was noticed because the
"options.resolvehostnames" test was failing on OS X when using Python 3
(it was showing an empty string in the output instead of the expected
string "<???>").

Also fixed the check to detect if Python 3 is running (in case the major
version is incremented in the future, the "using_py3" variable should
be True).