Various minor code cleanup (no changes in functionality)
Use "with" statement to open files so that we don't need to remember
to close them.  Remove an unused variable and import.  Remove unneeded
calls to dictionary keys() method.  Whitespace readability improvements.
Replace "X and Y or Z" with a conditional expression.  Add a few source
code comments.  Replace a "while" loop with code that doesn't need a
loop.  Fix some bare "except:" clauses to catch a specific exception
type.  Move a couple of checks for IP addr in local nets earlier in
the code to avoid unnecessary computation.