Cleanup output and make output easier to customize
Streamlined the output by removing two columns (the "Services" column
and the "State" column) when trace-summary is reading from a packet
trace, because ipsumdump doesn't provide that info.

Removed trailing whitespace from lines, and changed the default width
of the src and dst columns to be slightly narrower.

Changed the format() method to make it easier to customize the
output format.  The individual column widths, the left margin string,
and the delimiter string are now defined as variables.

Rearranged several lines of code to avoid unneeded computation.

Updated test baselines.